backend/quant/requirements.txt

ccxt==4.5.40
pandas==2.2.3
numpy==2.1.2
pyyaml==6.0.2
tenacity==9.0.0
python-dateutil==2.9.0.post0
fastapi==0.134.0
starlette==0.49.1
aiohttp==3.13.3
uvicorn==0.35.0
httpx==0.28.1
